requirements/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
twilio==9.3.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.13.0
